Add files via upload
Resume screening is the process of evaluating and shortlisting job applicants based on their resumes. Traditionally, this process has been performed manually by recruiters and HR teams, which can be time-consuming and subject to bias. However, with the help of machine learning and natural language processing techniques, it is now possible to automate the initial stages of the recruitment process.

A resume screening project typically involves developing a model that can extract relevant information from resumes, such as skills, experience, education, and more. This model can be trained on a large dataset of resumes, which enables it to learn to identify patterns and make predictions about which candidates are most likely to be a good fit for the job position.

By resume screening, recruiters and HR teams can save valuable time and resources, and ensure that they are selecting the most qualified candidates for the job. Additionally, machine learning-based resume screening can help reduce bias in the recruitment process by removing human judgement and relying on data-driven decision making.